// src/types/storage.ts
export enum SecureStorageKey {
    AuthenticationMethod = "AuthenticationMethod",
    RefreshToken = "RefreshToken",
    AccessToken = "AccessToken",
    PendingSession = "PendingSession",
    AuthenticationPin = "AuthenticationPin",
    RemainingAuthenticationPinAttempts = "RemainingAuthenticationPinAttempts",
}